create opcode_handlers folder and move arithmetic opcodes
